feat: create order

+++ src/main/java/com/example/farmfarm_refact/entity/OrderDetailEntity.java
@@ -22,7 +22,7 @@ public class OrderDetailEntity {
     @Id
     @GeneratedValue
     @Column(name="od_id")
-    private Long odId;
+    private Integer odId;
 
     private int quantity;
